refactor: migrate 'upload' commands
