+++ tests/unit/commands/test_upload.py
@@ -79,7 +79,7 @@ def snap_file_with_started_at():
 @pytest.mark.usefixtures("memory_keyring")
 @pytest.mark.parametrize(
     "command_class",
-    (commands.core22.StoreUploadCommand, commands.legacy.StoreLegacyPushCommand),
+    (commands.StoreUploadCommand, commands.legacy.StoreLegacyPushCommand),
 )
 def test_default(
     emitter,
@@ -114,7 +114,7 @@ def test_default(
 @pytest.mark.usefixtures("memory_keyring")
 @pytest.mark.parametrize(
     "command_class",
-    (commands.core22.StoreUploadCommand, commands.legacy.StoreLegacyPushCommand),
+    (commands.StoreUploadCommand, commands.legacy.StoreLegacyPushCommand),
 )
 def test_built_at(
     emitter,
@@ -150,7 +150,7 @@ def test_built_at(
 def test_default_channels(
     emitter, fake_store_notify_upload, fake_store_verify_upload, snap_file
 ):
-    cmd = commands.core22.StoreUploadCommand(None)
+    cmd = commands.StoreUploadCommand(None)
 
     cmd.run(
         argparse.Namespace(
@@ -176,7 +176,7 @@ def test_default_channels(
 
 
 def test_invalid_file():
-    cmd = commands.core22.StoreUploadCommand(None)
+    cmd = commands.StoreUploadCommand(None)
 
     with pytest.raises(craft_cli.errors.ArgumentParsingError) as raised:
         cmd.run(
